refactor(backend): extract shared Deezer cover URL helper
